==> ATAC/aggregated_analysis/00_update_metadata.ipynb <==
### Reformat original metadata

- `00_original_metadata.csv` contains metadata information from individually examining each study. We will reformat this to make sure that its metadata is concordant with the snRNA-seq aggregated analysis, especially with Multiome data

import pandas as pd
import numpy as np
import scanpy as sc
from collections import Counter
import os

metadata_df = pd.read_csv("00_original_metadata.csv", index_col = 0)
metadata_df = metadata_df[["sample_id", "donor_id", "study", "age_status", "age", 
                           "sex", "region", "disease_binary", "technology", "full_path"]].reset_index(drop=True)
metadata_df.head()

metadata_df['fragment_file'] = metadata_df['full_path'].str.split("/").str[-1]
metadata_df = metadata_df.drop(columns = "full_path")
metadata_df['full_path'] = metadata_df['fragment_file']

Counter(metadata_df.study)

#### Add the directory to fragment file name to get the full path

study_to_base_path = {
    "Kuppe": "/mnt/data1/william/human_heart_project/Final_manuscript_analysis/ATAC/external_datasets/Kuppe/",
    "ENCODE": "/mnt/data1/william/human_heart_project/Final_manuscript_analysis/ATAC/external_datasets/ENCODE/",
    "Penn": "/mnt/data1/william/human_heart_project/Final_manuscript_analysis/ATAC/internal_datasets/fragment_files/",
    "Kanemaru": "/mnt/data1/william/human_heart_project/Final_manuscript_analysis/ATAC/external_datasets/Kanemaru/",
    "Ameen": "/mnt/data1/william/human_heart_project/Final_manuscript_analysis/ATAC/external_datasets/Ameen/",
}

for study, base_path in study_to_base_path.items():
    mask = metadata_df['study'] == study  # Create a mask for the current study
    metadata_df.loc[mask, 'full_path'] = base_path + metadata_df.loc[mask, 'full_path']

# check that file exists
metadata_df['file_exists'] = metadata_df['full_path'].apply(os.path.exists)

print(metadata_df[metadata_df['file_exists'] != True])

#### Check the metadata and make it consistent with snRNA-seq

Counter(metadata_df.region)

Counter(metadata_df.sample_id)

Counter(metadata_df.donor_id)

Counter(metadata_df.age_status)

Counter(metadata_df.disease_binary)

metadata_df.loc[metadata_df.disease_binary == "non-diseased", 'disease_binary'] = "N"
metadata_df.loc[metadata_df.disease_binary == "diseased", 'disease_binary'] = "Y"
Counter(metadata_df.disease_binary)

Counter(metadata_df.sex)

### Load in the RNA adata

%%time
RNA_adata = sc.read_h5ad("../../RNA/aggregated_analysis/03_combined_all_snRNA.h5ad")
RNA_adata

RNA_donors = set(RNA_adata.obs.donor_id)
RNA_donors

### The ATAC donor_id and studies need to be consistent.

#### For study, we will add the year of the study

Counter(metadata_df['study'])

metadata_df['revised_study'] = metadata_df['study'].map(
    {
        "Kuppe": "Kuppe 2022",
        "Kanemaru": "Kanemaru 2023",
        "ENCODE": "ENCODE v4 (Snyder)",
        "Penn": "Penn",
        "Ameen": "Ameen 2022"
    }
)

metadata_df['study'] = metadata_df['revised_study']
metadata_df = metadata_df.drop(columns = ['revised_study'])

Counter(metadata_df['study'])

#### Now, add the study to the donor_id in the ATAC metadata

metadata_df['donor_id'] = metadata_df['study'] + ":" + metadata_df['donor_id']

# examine intersection between RNA and ATAC 
set(metadata_df['donor_id']) & RNA_donors

metadata_df[metadata_df['study']  == "Kanemaru 2023"]

#### The Kanemaru donor_ids need to be renamed to be consistent with the names in the snRNA-seq dataset
RNA-seq: 
- 'Kanemaru 2023:AH1-Nuclei_Multiome-v1'
- 'Kanemaru 2023:D3-Nuclei_Multiome-v1'
- 'Kanemaru 2023:D7-Nuclei_Multiome-v1'
- 'Kanemaru 2023:D8-Nuclei_Multiome-v1',

currently ATAC-seq is just: Kanemaru 2023:D3, Kanemaru 2023:D7, Kanemaru 2023:D8, Kanemaru 2023:AH1

metadata_df.loc[metadata_df['donor_id'] == "Kanemaru 2023:D3", 'donor_id'] = "Kanemaru 2023:D3-Nuclei_Multiome-v1"
metadata_df.loc[metadata_df['donor_id'] == "Kanemaru 2023:D7", 'donor_id'] = "Kanemaru 2023:D7-Nuclei_Multiome-v1"
metadata_df.loc[metadata_df['donor_id'] == "Kanemaru 2023:D8", 'donor_id'] = "Kanemaru 2023:D8-Nuclei_Multiome-v1"
metadata_df.loc[metadata_df['donor_id'] == "Kanemaru 2023:AH1", 'donor_id'] = "Kanemaru 2023:AH1-Nuclei_Multiome-v1"

# examine intersection between RNA and ATAC 
set(metadata_df['donor_id']) & RNA_donors

#### It will be easier later if we use the ENCODE mapping file to change the names of the ENCODE donor_ids to their corresponding RNA donor_id

ENCODE_mapping_df = pd.read_csv("multiome_mapping_files/ENCODE_ATAC_RNA_mapping.txt", delim_whitespace=True)
ENCODE_mapping_df.head()

# add the study name so that the donor_ids can be mapped to each other
ENCODE_mapping_df['RNA_directory_name'] = "ENCODE v4 (Snyder):" + ENCODE_mapping_df['RNA_directory_name']
ENCODE_mapping_df['ATAC_directory_name'] = "ENCODE v4 (Snyder):" + ENCODE_mapping_df['ATAC_directory_name']

Now, change the ATAC metadata_df's donor_id to the RNA_directory_name

merged_metadata_df = metadata_df.merge(
    ENCODE_mapping_df[['ATAC_directory_name', 'RNA_directory_name']],  
    left_on='donor_id',                                         
    right_on='ATAC_directory_name',                             
    how='left'                                                  
)

# update donor_id with 'RNA_directory_name' if 'RNA_directory_name' is not NaN
merged_metadata_df['donor_id'] = merged_metadata_df['RNA_directory_name'].where(
    merged_metadata_df['RNA_directory_name'].notna(),  
    merged_metadata_df['donor_id']                     
)

# drop merged df columns no longer needed
merged_metadata_df = merged_metadata_df.drop(columns = ["ATAC_directory_name", "RNA_directory_name"])

# now examine the intersection
set(merged_metadata_df['donor_id']) & RNA_donors

set(merged_metadata_df['donor_id']) - RNA_donors

### The non-intersecting donors are all accounted for:

- Ameen 2022 only performed snATAC-seq
- Penn F3 to F5 did not have corresponding RNA libraries
- ENCFF069ATM did not have very many high quality snRNA-seq nuclei after QC filtering
- There are 8 other ENCODE libraries for which only snATAC-seq was performed; they were NOT part of the Multiomics series
- ENCSR409QLZ is part of the multiomic series, but the Multiomics Series: ENCSR519BTE is accessed denied
- Kuppe P5 is the CAD,HF patient which we dropped as we cannot ascertain the patient's HF status

#### Based on this, we will drop the following patients that from the analysis as they lack a snRNA-seq correspondence or that snRNA-seq correspondence was removed: ENCODE v4 (Snyder):ENCSR409QLZ, Kuppe 2022:P5

#### Before doing this, indicate the number of donors per study 

merged_metadata_df[["study", "donor_id"]].drop_duplicates().groupby("study").count()

Remove those 2 samples

remove_samples = ["ENCODE v4 (Snyder):ENCSR409QLZ", "Kuppe 2022:P5"]

filt_metadata_df = merged_metadata_df[~merged_metadata_df.donor_id.isin(remove_samples)].reset_index(drop=True).copy()

# drop the file_exists column
filt_metadata_df = filt_metadata_df.drop(columns = ["file_exists"])

len(set(filt_metadata_df.donor_id) & RNA_donors)

len(set(filt_metadata_df.donor_id) - RNA_donors)

len(RNA_donors - set(filt_metadata_df.donor_id))

merged_metadata_df[["study", "technology", "donor_id"]].drop_duplicates().groupby(["study", "technology"]).count()

#### Therefore, there are 200 donors for which there is only RNA, 91 for which there are both RNA + ATAC, and 15 for which there are only ATAC

#RNA_donors - set(filt_metadata_df.donor_id)

### Finally, inspect each study's metadata carefully one more time to confirm information

merged_metadata_df[merged_metadata_df['study'] == "Penn"].sort_values(by = "donor_id")

merged_metadata_df[merged_metadata_df['study'] == "Kuppe 2022"].sort_values(by = "donor_id")

merged_metadata_df[merged_metadata_df['study'] == "Ameen 2022"].sort_values(by = "donor_id")

merged_metadata_df[merged_metadata_df['study'] == "Kanemaru 2023"].sort_values(by = "donor_id")

merged_metadata_df[merged_metadata_df['study'] == "ENCODE v4 (Snyder)"].sort_values(by = "donor_id")

### Save this updated metadata

filt_metadata_df.head()

filt_metadata_df.to_csv("00_updated_metadata.csv")
