==> ATAC/aggregated_analysis/01C_update_metadata_with_QC.ipynb <==
#### Add QC filters for all samples, which is located in `01B_QC_filter_per_sample.txt` into the metadata

import pandas as pd
import numpy as np
import scanpy as sc
from collections import Counter
import os

metadata_df = pd.read_csv("00_updated_metadata.csv", index_col = 0)
metadata_df.shape

metadata_df.head()

### load in the QC information
QC_filters_df = pd.read_csv("01B_QC_filter_per_sample.txt", index_col = 0, delim_whitespace=True).reset_index()
QC_filters_df.head()

QC_filters_df['sample_id'] = QC_filters_df['file'].str.replace("_pre_trimmed_tsse_v_nfrags.png", "")

merged_df = metadata_df.merge(QC_filters_df, on = "sample_id")

len ( merged_df.donor_id.unique() )

merged_df.to_csv("01C_QC_updated_metadata.csv")
