==> ATAC/aggregated_analysis/07A_make_ATAC_adata_consistent_with_RNA.ipynb <==
#### Make the metadata columns consistent between RNA and ATAC finalized objects. For cases where there are any inconsistencies in `06D_filtered_peak_matrix.h5ad` (ATAC object), we will make it cohere with the RNA finalized object (`07_final_RNA_without_scvi.h5ad`)

import scanpy as sc 
import snapatac2 as snap
import pandas as pd
import numpy as np
from collections import Counter

%%time
ATAC_adata = sc.read_h5ad("06D_filtered_peak_matrix.h5ad")
ATAC_adata

%%time
RNA_adata = sc.read_h5ad("../../RNA/aggregated_analysis/07_final_RNA_without_scvi.h5ad")
RNA_adata

### Check the columns

ATAC_metadata_columns = list(ATAC_adata.obs.columns)
RNA_metadata_columns = list(RNA_adata.obs.columns)

set(ATAC_metadata_columns) - set(RNA_metadata_columns)

set(RNA_metadata_columns) - set(ATAC_metadata_columns)

set(RNA_metadata_columns) & set(ATAC_metadata_columns)

#### First, we will add any metadata missing in the ATAC_adata that is present in the RNA adata and vice versa:

- Missing in the RNA adata: none, since the columns in the ATAC only are specific to that assay
- Missing in the ATAC adata: cell_or_nuclei (all nuclei); age_group, decade, and disease (AMI for the ATAC), final_cell_type

#### Then we will make sure that the naming scheme is consistent for the ATAC and RNA for the shared columns

def get_age_group(age):
    '''
    Add age group to categorize age into three discrete categories based on the decade of donor
    '''
    if age < 40:
        age_group = "young"
    elif (age >= 40) & (age < 60):
        age_group = "middle"
    elif (age >= 60):
        age_group = "old"
    return(age_group)

%%time
# add age group and decade
ATAC_adata.obs['age_group'] = ATAC_adata.obs['age'].apply(lambda x: get_age_group(x))
# set fetal as an age group
ATAC_adata.obs.loc[ATAC_adata.obs.age_status == "Fetal", 'age_group'] = "fetal"

ATAC_adata.obs['decade'] = np.floor(ATAC_adata.obs['age'] / 10)

ATAC_adata.obs['final_cell_type'] = ATAC_adata.obs['cell_type']

ATAC_adata.obs['cell_or_nuclei'] = "Nuclei"

Counter(RNA_adata.obs.disease)

Counter(ATAC_adata.obs.disease_binary)

# add disease
ATAC_adata.obs['disease'] = "ND"
ATAC_adata.obs.loc[ATAC_adata.obs.disease_binary == "Y", 'disease'] = "AMI"

Now check the ATAC and RNA shared columns

ATAC_metadata_columns = list(ATAC_adata.obs.columns)
RNA_metadata_columns = list(RNA_adata.obs.columns)

set(RNA_metadata_columns) - set(ATAC_metadata_columns)

shared_metadata_columns = list( set(RNA_metadata_columns) & set(ATAC_metadata_columns) )

for metadata_col in shared_metadata_columns:
    if not metadata_col in ["barcode", "cell_type", "sample_id", "donor_id"]:
        print(metadata_col)
        
        print("only in ATAC:")
        print(set(ATAC_adata.obs[metadata_col]) - set(RNA_adata.obs[metadata_col]))
        
        print("only in RNA:")
        print(set(RNA_adata.obs[metadata_col]) - set(ATAC_adata.obs[metadata_col]))
        
        print("shared:") 
        print(set(RNA_adata.obs[metadata_col]) & set(ATAC_adata.obs[metadata_col]))
        
        print("\n")

### Make the metadata consistent

- age_status: make Postnatal and Fetal in ATAC lowercase
- technology, tech_plus_study: investigate why some of them are nan and fix this

%%time

# make age status lowercase
ATAC_adata.obs['age_status'] = ATAC_adata.obs['age_status'].str.lower()

Counter(ATAC_adata.obs.study)

Counter(ATAC_adata.obs.technology)

# reassign the technology based on the studies
ATAC_adata.obs['technology'] = ATAC_adata.obs['study'].map(
    {
        "ENCODE v4 (Snyder)": "Multiome-v1", 
        "Penn": "10X_ATAC",
        "Kuppe 2022": "10X_ATAC",
        "Kanemaru 2023": "Multiome-v1",
        "Ameen 2022": "10X_ATAC"
    }
)

Counter(ATAC_adata.obs.technology)

ATAC_adata.obs['tech_plus_study'] = ATAC_adata.obs['technology'].astype(str) + "_" + ATAC_adata.obs['study'].astype(str)

### Recheck the consistency of the metadata

shared_metadata_columns = list( set(RNA_metadata_columns) & set(ATAC_metadata_columns) )

for metadata_col in shared_metadata_columns:
    if not metadata_col in ["barcode", "cell_type", "sample_id", "donor_id"]:
        print(metadata_col)
        
        print("only in ATAC:")
        print(set(ATAC_adata.obs[metadata_col]) - set(RNA_adata.obs[metadata_col]))
        
        print("only in RNA:")
        print(set(RNA_adata.obs[metadata_col]) - set(ATAC_adata.obs[metadata_col]))
        
        print("shared:") 
        print(set(RNA_adata.obs[metadata_col]) & set(ATAC_adata.obs[metadata_col]))
        
        print("\n")

### Now that the metadata is consistent with the RNA adata, save this final ATAC adata

ATAC_adata

%%time
ATAC_adata.write("07_final_ATAC.h5ad")
