==> ATAC/aggregated_analysis/07B_get_peak_bed.ipynb <==
import pandas as pd
import numpy as np
import scanpy as sc

%%time
peak_mat = sc.read_h5ad("07_final_ATAC.h5ad")
peak_mat

def extract_genome_coordinates(peak_df, column_name):
    '''Convert a column of a pandas df that is of the format (chr:start-end) to three different columns, 
    called chr, start, and end. Specify the column_name for original column.'''

    peak_df[["chr", "coords"]] =  peak_df[column_name].str.split(":", expand=True)
    peak_df[["start", "end"]] = peak_df["coords"].str.split("-", expand=True)
    peak_df = peak_df.drop(columns=["coords"])

    return peak_df

peak_names_df = peak_mat.var
peak_names_df = peak_names_df.reset_index().rename(columns = {'index': 'coordinates'})
peak_names_df = extract_genome_coordinates(peak_names_df, column_name = "coordinates")

peak_names_df.shape

peak_names_df.head()

### Save in a BED format file

peak_names_bed = peak_names_df[["chr", "start", "end"]]
peak_names_bed.to_csv("07B_snATAC_peaks.bed", header=False, index=None, sep = "\t")
